big-refactor
:heavy_check_mark: adding features to maf concat
:heavy_check_mark: big restructure so annotate is under maf command
:heavy_check_mark: update maf concat tests
