+++ postprocessing_variant_calls/maf/concat/concat_process.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+# We can use this file if we ever want concat to have sub-commands,
+# otherwise concat lives in main.py of the maf folder
