+++ resources/header.txt
@@ -1,0 +1,1 @@
+Hugo_Symbol,Chromosome,Start_Position,End_Position,Reference_Allele,Tumor_Seq_Allele2,Variant_Classification,Variant_Type,Tumor_Sample_Barcode,Matched_Norm_Sample_Barcode,HGVSp_Short,caller_t_ref_count,caller_t_alt_count,caller_n_ref_count,caller_n_alt_count
