+++ tests/data/maf/concat/maf1.maf
@@ -1,0 +1,3 @@
+Hugo_Symbol	Chromosome	Start_Position	End_Position	Reference_Allele	Tumor_Seq_Allele2	Variant_Classification	Variant_Type	Tumor_Sample_Barcode	Matched_Norm_Sample_Barcode	HGVSp_Short	caller_t_ref_count	caller_t_alt_count	caller_n_ref_count	caller_n_alt_count
+LOC101928728	1	27022823	27022824	-	G	Frame_Shift_Ins	INS	C-1234	RU776684-N	p.R25Pfs*58	103	3	332	2
+LOC101928728	1	27022837	27022837	A	-	Frame_Shift_Del	DEL	C-1234	RU776684-N	p.T22Pfs*63	83	1	314	0
